// generated/api-client/models/events-query-controller-get-event-attendees200-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { EventsQueryControllerGetEventAttendees200ResponseAttendeesInner } from './events-query-controller-get-event-attendees200-response-attendees-inner';
// May contain unused imports in some cases
// @ts-ignore
import { EventsQueryControllerGetEventAttendees200ResponseMeta } from './events-query-controller-get-event-attendees200-response-meta';

/**
 * 
 * @export
 * @interface EventsQueryControllerGetEventAttendees200Response
 */
export interface EventsQueryControllerGetEventAttendees200Response {
    /**
     * 
     * @type {string}
     * @memberof EventsQueryControllerGetEventAttendees200Response
     */
    'message'?: string;
    /**
     * 
     * @type {number}
     * @memberof EventsQueryControllerGetEventAttendees200Response
     */
    'statusCode'?: number;
    /**
     * 
     * @type {Array<EventsQueryControllerGetEventAttendees200ResponseAttendeesInner>}
     * @memberof EventsQueryControllerGetEventAttendees200Response
     */
    'attendees'?: Array<EventsQueryControllerGetEventAttendees200ResponseAttendeesInner>;
    /**
     * 
     * @type {EventsQueryControllerGetEventAttendees200ResponseMeta}
     * @memberof EventsQueryControllerGetEventAttendees200Response
     */
    'meta'?: EventsQueryControllerGetEventAttendees200ResponseMeta;
}