// generated/api-client/models/events-query-controller-get-event-attendees200-response-attendees-inner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface EventsQueryControllerGetEventAttendees200ResponseAttendeesInner
 */
export interface EventsQueryControllerGetEventAttendees200ResponseAttendeesInner {
    /**
     * 
     * @type {string}
     * @memberof EventsQueryControllerGetEventAttendees200ResponseAttendeesInner
     */
    'username'?: string;
    /**
     * 
     * @type {string}
     * @memberof EventsQueryControllerGetEventAttendees200ResponseAttendeesInner
     */
    'firstName'?: string;
    /**
     * 
     * @type {string}
     * @memberof EventsQueryControllerGetEventAttendees200ResponseAttendeesInner
     */
    'lastName'?: string;
    /**
     * 
     * @type {string}
     * @memberof EventsQueryControllerGetEventAttendees200ResponseAttendeesInner
     */
    'imageUrl'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EventsQueryControllerGetEventAttendees200ResponseAttendeesInner
     */
    'email'?: string;
    /**
     * 
     * @type {number}
     * @memberof EventsQueryControllerGetEventAttendees200ResponseAttendeesInner
     */
    'totalDuration'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof EventsQueryControllerGetEventAttendees200ResponseAttendeesInner
     */
    'isEligible'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof EventsQueryControllerGetEventAttendees200ResponseAttendeesInner
     */
    'registeredAt'?: string;
    /**
     * 
     * @type {number}
     * @memberof EventsQueryControllerGetEventAttendees200ResponseAttendeesInner
     */
    'sessionCount'?: number;
}